Updating task specifications, company can see all its task and can also review submissions

+++ src/main/java/com/chanderparkash/internx/Entities/Rating.java
@@ -1,5 +1,9 @@
 package com.chanderparkash.internx.Entities;
 
+import java.time.LocalDateTime;
+
+import org.hibernate.annotations.CreationTimestamp;
+
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -11,8 +15,6 @@ import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import java.time.LocalDateTime;
-import org.hibernate.annotations.CreationTimestamp;
 
 @Entity
 @Data
@@ -36,7 +38,7 @@ public class Rating {
     @JoinColumn(name = "task_id")
     private Tasks task;
 
-    private Integer rating;
+    private Double rating;
     @Column(columnDefinition = "TEXT")
     private String feedback;
     @CreationTimestamp
